Add files via upload
When you have a large file that you want to divide into two or more parts, you'll need a software program to carry out the process. That's where the option of using a PDF splitter comes in. With this tool, you can separate individual or multiple pages from the original document without destroying it using python